script-data/requirements.txt
numpy>=1.26
pandas>=2.2.2
python-dotenv>=1.0.0
sqlalchemy>=1.4.39 
pyodbc>=4.0.35
requests>=2.29.0
pymssql==2.2.10
influxdb==5.3.1
ipykernel==6.26.0
streamlit==1.27.0
paho-mqtt==1.6.1
altair ==5.4.1